+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/feature/SpawnerMobsExperience.java
@@ -14,13 +14,16 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 public class SpawnerMobsExperience extends Feature {
 	private final ForgeConfigSpec.ConfigValue<Double> mobsFromSpawnersMultiplierConfig;
 
-	public double mobsFromSpawnersMultiplier = 0.667d;
+	public double mobsFromSpawnersMultiplier = 0.5d;
 
 	public SpawnerMobsExperience(Module module) {
 		super(Config.builder, module, true);
 		Config.builder.comment(this.getDescription()).push(this.getName());
 		mobsFromSpawnersMultiplierConfig = Config.builder
-				.comment("Experience dropped from mobs that come from spawners will be multiplied by this multiplier. Experience dropped by mobs from spawners are still affected by 'Global Experience Multiplier'\nCan be set to 0 to disable experience drop from mob that come from spawners.")
+				.comment("""
+						Experience dropped from mobs that come from spawners will be multiplied by this multiplier.
+						Experience dropped by mobs from spawners are still affected by 'Global Experience Multiplier'
+						Can be set to 0 to disable experience drop from mob that come from spawners.""")
 				.defineInRange("Mobs from Spawners Multiplier", this.mobsFromSpawnersMultiplier, 0.0d, 1000d);
 		Config.builder.pop();
 	}